requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-types==0.7.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
Deprecated==1.3.1
distro==1.9.0
frozenlist==1.8.0
greenlet==3.4.0
idna==3.11
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
netaddr==1.3.0
propcache==0.4.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
wrapt==2.1.2
yarl==1.23.0
yookassa==3.10.0
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.2
fastapi==0.136.0
frozenlist==1.8.0
greenlet==3.4.0
h11==0.16.0
idna==3.11
libpass==1.9.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.1
netaddr==1.3.0
propcache==0.4.1
pyasn1==0.6.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
requests==2.33.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
wrapt==2.1.2
yarl==1.23.0
yookassa==3.10.0
